requirements.txt
loguru

# Streamlit and web framework
streamlit>=1.28.0
streamlit-autorefresh>=0.0.1

# Data fetching and APIs
yfinance>=0.2.18
requests>=2.31.0

# Data analysis and portfolio optimization
pandas>=2.0.3
numpy>=1.24.3
scipy>=1.11.1
scikit-learn>=1.3.0
cvxpy>=1.3.2

# Visualization
plotly>=5.15.0
matplotlib>=3.7.2
seaborn>=0.12.2

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3

# Optional: For advanced portfolio optimization
pypfopt>=1.5.4

# Optional: For sentiment analysis (if implementing news sentiment)
textblob>=0.17.1
vaderSentiment>=3.3.2

# Testing (optional for development)
pytest>=7.4.0
pytest-mock>=3.11.1

# Logging
loguru>=0.7.0